+++ src/main/scala/edg_ide/util/StreamUtils.scala
@@ -1,0 +1,15 @@
+package edg_ide.util
+
+import java.io.InputStream
+
+object StreamUtils {
+  def forAvailable(stream: InputStream)(fn: Array[Byte] => Unit): Unit = {
+    var available = stream.available()
+    while (available > 0) {
+      val array = new Array[Byte](available)
+      stream.read(array)
+      fn(array)
+      available = stream.available()
+    }
+  }
+}
